emacs: get rid of melpa-stable
Not sure how much this[^1] is accurate or current, but I realized that by setting melpa to the
highest priority I was never using melpa-stable.

The TL;DR of that post is the versions of packages on melpa-stable are not "stable" because of some
guarantee (like Debian Stable for example); instead they are just versions that are tagged... and a
lot of packages don't tag release with a good cadence.

[^1]: https://www.reddit.com/r/emacs/comments/etikbz/speaking_as_a_package_maintainer_please_do_not/