emacs: turn on flyspell for commits and set line length
When I noticed that the git-commit package was deprecated, I didn't
think of figuring out how to convert the settings that I was using,
notably the hook to turn on flyspell for commit messages and the
settings for subject and body length.

This [article](https://robertcooper.me/post/git-commit-messages) makes
an argument for subject and body length when using Github, and I think
it's reasonable enough.